Move validate_transaction to SignedTransaction
