+++ radix-engine/src/transaction/mod.rs
@@ -3,11 +3,9 @@ mod builder;
 mod error;
 mod executor;
 mod nonce_provider;
-mod validator;
 
 pub use abi_provider::{AbiProvider, BasicAbiProvider};
 pub use builder::TransactionBuilder;
 pub use error::{BuildArgsError, CallWithAbiError};
 pub use executor::TransactionExecutor;
 pub use nonce_provider::NonceProvider;
-pub use validator::validate_transaction;
